+++ code/src/sixsq/nuvla/server/resources/callback_2fa_activation.clj
@@ -1,12 +1,12 @@
 (ns sixsq.nuvla.server.resources.callback-2fa-activation
   "
-Allow a user to activate or deactivate two factor authentication.
+Allow a user to activate two factor authentication.
 "
   (:require
-    [clojure.string :as str]
     [clojure.tools.logging :as log]
     [sixsq.nuvla.server.resources.callback :as callback]
     [sixsq.nuvla.server.resources.callback.utils :as utils]
+    [sixsq.nuvla.server.resources.two-factor-auth.utils :as auth-2fa]
     [sixsq.nuvla.server.resources.user.utils :as user-utils]
     [sixsq.nuvla.server.util.log :as logu]
     [sixsq.nuvla.server.util.response :as r]))
@@ -14,36 +14,28 @@ Allow a user to activate or deactivate two factor authentication.
 
 (def ^:const action-name "2fa-activation")
 
-(def ^:const msg-wrong-2fa-token "Wrong 2FA token")
-
 (def create-callback (partial callback/create action-name))
-
-
-(defmulti token-is-valid? (fn [_request callback] (-> callback :data :method)))
-
-(defmethod token-is-valid? :default
-  [{{user-token :token} :body :as _request}
-   {{:keys [token]} :data :as _callback}]
-  (and
-    (not (str/blank? user-token))
-    (= user-token token)))
-
 
 (defmethod callback/execute action-name
   [{{user-id :href}         :target-resource
-    {:keys [method enable]} :data
+    {:keys [method secret]} :data
     callback-id             :id
     :as                     callback}
    request]
   (try
     (utils/callback-dec-tries callback-id)
-    (if (token-is-valid? request callback)
-      (let [msg (str "2FA with method '" method "' " (if enable "activated" "disabled")
-                     " for " user-id ". Callback successfully executed.")]
-        (user-utils/update-user user-id {:auth-method-2fa method})
+    (if (auth-2fa/is-valid-token? method request callback)
+      (let [msg     (str "2FA with method '" method "' activated for " user-id
+                         ". Callback successfully executed.")
+            cred-id (when secret
+                      (user-utils/create-totp-credential user-id secret))]
+        (user-utils/update-user
+          user-id (cond-> {:auth-method-2fa method}
+                          cred-id (assoc :credential-totp cred-id)))
         (log/info msg)
         (utils/callback-succeeded! callback-id)
         (r/map-response msg 200 user-id))
-      (logu/log-and-throw-400 (str msg-wrong-2fa-token " for " user-id)))
+      (logu/log-and-throw-400
+        (str auth-2fa/msg-wrong-2fa-token " for " user-id)))
     (catch Exception e
       (or (ex-data e) (throw e)))))